Moves mpi client stub in its own module
