+++ dsf-bpe/pom.xml
@@ -164,6 +164,11 @@
 				<artifactId>dsf-mpi-client-pdq</artifactId>
 				<version>${project.version}</version>
 			</dependency>
+			<dependency>
+				<groupId>org.highmed.dsf</groupId>
+				<artifactId>dsf-mpi-client-stub</artifactId>
+				<version>${project.version}</version>
+			</dependency>
 
 			<dependency>
 				<groupId>org.camunda.bpm</groupId>
