+++ dsf-mpi/dsf-mpi-client-stub/pom.xml
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+		 xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+		 xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+	<modelVersion>4.0.0</modelVersion>
+
+	<artifactId>dsf-mpi-client-stub</artifactId>
+
+	<parent>
+		<artifactId>dsf-mpi-pom</artifactId>
+		<groupId>org.highmed.dsf</groupId>
+		<version>0.2.0-SNAPSHOT</version>
+	</parent>
+
+	<dependencies>
+		<dependency>
+			<groupId>org.highmed.dsf</groupId>
+			<artifactId>dsf-mpi-client</artifactId>
+		</dependency>
+	</dependencies>
+</project>
